test: move test_idf_version_keywords_type from test_build.py to test_manifest.py
