+++ tests/test_build.py
@@ -95,16 +95,3 @@ class TestBuild:
         for app in apps:
             app.build()
             assert app.build_status == BuildStatus.SUCCESS
-
-
-@pytest.mark.skipif(not shutil.which('idf.py'), reason='idf.py not found')
-def test_idf_version_keywords_type():
-    from idf_build_apps.constants import (
-        IDF_VERSION_MAJOR,
-        IDF_VERSION_MINOR,
-        IDF_VERSION_PATCH,
-    )
-
-    assert isinstance(IDF_VERSION_MAJOR, int)
-    assert isinstance(IDF_VERSION_MINOR, int)
-    assert isinstance(IDF_VERSION_PATCH, int)
